+++ CHANGES.md
@@ -4,6 +4,7 @@
 
 * `bijection-guava` for Guava interop.
 * `bijection-algebird` for bijections on classes in twitter's [Algebird](https://github.com/twitter/algebird).
+* `bijection-util` with bijections on [twitter-util](https://github.com/twitter/util)'s "Try" and "Future".
 
 ### 0.2.0